Move UtUtils Remove old example Do refactoring
