+++ src/test/java/com/github/dakusui/floorplan/examples/bookstore/components/Nginx.java
@@ -6,10 +6,8 @@ import com.github.dakusui.floorplan.resolver.Resolvers;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
-import static com.github.dakusui.floorplan.UtUtils.printf;
+import static com.github.dakusui.floorplan.ut.UtUtils.printf;
 import static com.github.dakusui.floorplan.resolver.Resolvers.*;
 
 public class Nginx {
