+++ src/test/java/com/github/dakusui/floorplan/examples/bookstore/components/PostgreSQL.java
@@ -4,7 +4,7 @@ import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.ComponentSpec;
 import com.github.dakusui.floorplan.component.Operator;
 
-import static com.github.dakusui.floorplan.UtUtils.printf;
+import static com.github.dakusui.floorplan.ut.UtUtils.printf;
 import static com.github.dakusui.floorplan.resolver.Resolvers.immediate;
 import static com.github.dakusui.floorplan.resolver.Resolvers.slotValue;
