+++ src/test/java/com/github/dakusui/floorplan/ut/ListReferenceTest.java
@@ -1,7 +1,6 @@
 package com.github.dakusui.floorplan.ut;
 
 import com.github.dakusui.floorplan.Fixture;
-import com.github.dakusui.floorplan.UtUtils;
 import com.github.dakusui.floorplan.component.Attribute;
 import com.github.dakusui.floorplan.component.Component;
 import com.github.dakusui.floorplan.component.ComponentSpec;
